Finished the template flow, only attributes and seeding left before kafka
