+++ InvTemplateInfo/functionalities/invtemplate/InvTemplateController.cs
@@ -1,4 +1,5 @@
 ﻿using Dependencies;
+using InvTemplateInfo.functionalities.invtemplate.dtos;
 using InvTemplateInfo.functionalities.invtemplate.repo;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
@@ -27,6 +28,14 @@ namespace InvTemplateInfo.functionalities.invtemplate
             endpoints.MapGet("/invtemplate/allpermissions", GetPermissions)
                 .RequireAuthorization(p => p.RequireClaim(ImportantStrings.Purpose, ImportantStrings.Access)
                                             .RequireClaim(ImportantStrings.PermissionSet, TemplatesStrings.TemplateAdmin));
+            
+            endpoints.MapPost("/invtemplate", SaveTemplate)
+                .RequireAuthorization(p => p.RequireClaim(ImportantStrings.Purpose, ImportantStrings.Access)
+                                            .RequireClaim(ImportantStrings.PermissionSet, TemplatesStrings.TemplateAdmin));
+            endpoints.MapPost("/invtemplate/release", ReleaseTemplate)
+                .RequireAuthorization(p => p.RequireClaim(ImportantStrings.Purpose, ImportantStrings.Access)
+                                            .RequireClaim(ImportantStrings.PermissionSet, TemplatesStrings.TemplateAdmin));
+
         }
 
         public static async Task<IResult> GetTemplates([FromServices] PTemplateRepo templateRepo)
@@ -44,6 +53,25 @@ namespace InvTemplateInfo.functionalities.invtemplate
         public static async Task<IResult> GetPermissions([FromServices] PTemplateRepo templateRepo)
         {
             return TypedResults.Ok(await templateRepo.GetPermissions());
+        }
+
+        public static async Task<IResult> SaveTemplate([FromBody] TemplateForCreationDto dto, [FromServices] PTemplateRepo templateRepo)
+        {
+            if (await templateRepo.ExistsTemplate(dto.TemplateName, dto.TemplateVersion))
+                await templateRepo.DeleteAndRecreateParams(dto);
+            else
+                await templateRepo.CreateTemplate(dto);
+            return TypedResults.NoContent();
+        }
+
+        public static async Task<IResult> ReleaseTemplate([FromBody] TemplateForReleaseDto dto, [FromServices] PTemplateRepo templateRepo)
+        {
+            if (await templateRepo.ExistsTemplate(dto.TemplateName, dto.TemplateVersion))
+            {
+                await templateRepo.ReleaseTemplate(dto);
+                return TypedResults.NoContent();
+            }
+            return TypedResults.BadRequest("Bad template");
         }
 
     }
