+++ InvTemplateInfo/functionalities/invtemplate/dtos/TemplateForReleaseDto.cs
@@ -1,0 +1,8 @@
+﻿namespace InvTemplateInfo.functionalities.invtemplate.dtos
+{
+    public struct TemplateForReleaseDto
+    {
+        public string TemplateName { get; set; }
+        public int TemplateVersion { get; set; }
+    }
+}
