+++ InvTemplateInfo/pages/invtemplate/template.css
@@ -1,6 +1,6 @@
-#quarters {
+#onetwo {
     display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
+    grid-template-columns: 1fr 2fr 1fr 2fr 1.5fr;
     justify-items: center;
     row-gap: 1em;
     column-gap: 1em;
