--- InvTemplateInfo/pages/testtemp/template.css
@@ -1,7 +1,0 @@
-#onetwo {
-    display: grid;
-    grid-template-columns: 1fr 2fr 1fr 2fr 1.5fr;
-    justify-items: center;
-    row-gap: 1em;
-    column-gap: 1em;
-}
