--- InvTemplateInfo/pages/testtemp/template.html
@@ -1,116 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <link rel="stylesheet" href="/public/color">
-    <link rel="stylesheet" href="/public/font">
-    <link rel="stylesheet" href="/public/style">
-    <link rel="stylesheet" href="/public/topbar">
-    <script src="/public/material-components" type="module"></script>
-
-    <link rel="stylesheet" href="/public/centered-body">
-    <link rel="stylesheet" href="/invtemplate/test/style">
-    <script src="/public/authenticated-fetch" type="module"></script>
-    <script src="/invtemplate/test/script" type="module"></script>
-
-
-    <title>Groups</title>
-</head>
-
-<body>
-    <div id="topBar">
-        <md-filled-tonal-icon-button onclick="history.back();">
-            <img src="/public/back-logo" alt="Back arrow" />
-        </md-filled-tonal-icon-button>
-        <a href="/"><img src="/favicon.ico" alt="Modsig logo" /></a>
-        <md-filled-tonal-button onclick="location.href='/invtemplate/permission';">Permissions</md-filled-tonal-button>
-    </div>
-
-    <section>
-        <h3>Inventory Templates</h3>
-    </section>
-
-    <div id="onetwo">
-        <div class="verticalLine" id="templateContainer">
-            <section class="horizontalLine plusContainer">
-                <md-outlined-text-field class="name" label="Template Name" value="">
-                </md-outlined-text-field>
-
-                <md-filled-tonal-icon-button class="plusButton">
-                    <img src="/public/plus-logo" alt="Plus" />
-                </md-filled-tonal-icon-button>
-            </section>
-        </div>
-
-        <div class="verticalLine" id="attributeContainer">
-            <section class="horizontalLine">
-                <h3>Property1</h3>
-                <p>Script name</p>
-                <p>Actual edit</p>
-                <md-filled-tonal-icon-button class="deleteButton">
-                    <img src="/public/trashcan-logo" alt="Delete" />
-                </md-filled-tonal-icon-button>
-                <md-filled-tonal-icon-button class="editButton">
-                    <img src="/public/edit-logo" alt="Edit" />
-                </md-filled-tonal-icon-button>
-                <md-filled-tonal-icon-button class="resetButton">
-                    <img src="/public/reset-logo" alt="Reset" />
-                </md-filled-tonal-icon-button>
-            </section>
-
-            <section class="horizontalLine plusContainer">
-                <md-outlined-text-field class="name" label="Template Name" value="">
-                </md-outlined-text-field>
-
-                <md-outlined-select class="actionsContainer">
-                </md-outlined-select>
-
-                <md-filled-tonal-icon-button class="plusButton">
-                    <img src="/public/plus-logo" alt="Plus" />
-                </md-filled-tonal-icon-button>
-            </section>
-        </div>
-
-        <div class="verticalLine" id="attributePermissionsContainer">
-            
-        </div>
-
-        <div class="verticalLine" id="entityAttributeContainer">
-            <section class="horizontalLine">
-                <h3>Property1</h3>
-                <p>Script name</p>
-                <md-filled-tonal-icon-button class="deleteButton">
-                    <img src="/public/trashcan-logo" alt="Delete" />
-                </md-filled-tonal-icon-button>
-                <md-filled-tonal-icon-button class="editButton">
-                    <img src="/public/edit-logo" alt="Edit" />
-                </md-filled-tonal-icon-button>
-                <md-filled-tonal-icon-button class="resetButton">
-                    <img src="/public/reset-logo" alt="Reset" />
-                </md-filled-tonal-icon-button>
-            </section>
-
-            <section class="horizontalLine plusContainer">
-                <md-outlined-text-field class="name" label="Template Name" value="">
-                </md-outlined-text-field>
-
-                <md-outlined-select class="actionsContainer">
-                </md-outlined-select>
-
-                <md-filled-tonal-icon-button class="plusButton">
-                    <img src="/public/plus-logo" alt="Plus" />
-                </md-filled-tonal-icon-button>
-            </section>
-        </div>
-
-        <div class="verticalLine" id="entityAttributePermissionsContainer">
-            
-        </div>
-
-    </div>
-
-</body>
-
-</html>
